src/bitcoin-prover/requirements.txt
python-bitcoinlib==0.12.2
requests==2.32.5